Initial commit for squiddy module
For firebreak I wanted to try to write a module using test driven
development that interacts with the Trello API.

I want to create a very simple class that given a Trello card URL, will
create a checklist called "Pull Requests" on the card if the checklist
doesn't exist.

It will then add an item to the checklist.

It will also be able to mark an item as complete.

This is essentially the same behaviour as `github-trello-poster`[1],
something that I found infinitely useful and have sorely missed at
FutureLearn.
[1] https://github.com/emmabeynon/github-trello-poster