Add dotenv gem for managing ENV vars
Normally our ENV vars are provided via github when we call an action.
For testing, we’ll need dummy versions of these, so we’re using dotenv
to set them. 

In order to run the specs we need to create a test.env file. We’re adding
a template for it here.