Load git client in docker image
We need to copy our git_client.rb file so we can reference it in bubble_merge.rb.
Because git_client.rb lives in the lib/ folder, we can’t copy it directly since
the COPY command only looks in the current folder.

We’ve hard to create a hard link to the file inside this folder. This allows
us to copy it to our docker image.