Add state to PullRequest
The GitHub Actions environment does not provide the current state of the
pull request, even though you can trigger actions based upon this.

We could pass through this particular data, but since we're already
querying the API, we can do different things based upon the state of the
pull request as returned by the API.

Rather than know the exact keywords for the "state" of the pull request,
this adds two methods that return true or false.