Add initial runner
This adds a basic executable and Dockerfile to run action.

It's definitely worth noting that due to how the way GitHub Actions
mounts volumes, you are not able to use `bundle exec` in the same way
since the directory it'll stick you in will not have access to the
bundled gems.

This means we have to use the `--system` flag and install all the gems
to the system.