Push gem to GitHub Packages
After doing some testing, I've realised that the quickest way to run
this action is install it as a Gem, and leverage the quick
`ruby/ruby-setup` action that installs Ruby and can do bundler caching.

This adds the workflow to push the gem on `master`. To push a new
release, update the `VERSION` file, which will get read in as the
de-facto version.

I've added the double pipe at the end of `gem push` to stop builds
failing when that version already exists.