Add some rescue logic
I discovered a bug almost by accident when I closed a PR with a link to
a Trello card, where the checklist didn't exist.

The action failed, and I think in this particular case it should not
fail the action.

I do think if a card does not exist then it should raise an error; but
when it comes to ChecklistNotFound, it shouldn't.