Introduce GitClient to talk to the github API
This uses the octokit gem to perform actions related to our bubble merge
such as rebasing, commenting and merging.

It is using Squiddy's github token and parses the GITHUB_EVENT env var
sent to us when the action was invoked. This contains info such as the
branch name and the base branch.