Merge with rebase and adding a commit message
The merge should be performed after rebase on master and with the --no-ff
option in order to generate a merge bubble.

The merge bubble will be clean (only one bubble per segment) when this action
is combined with the rebase of the branch before merging. Currently we are
doing this by configuring the workflow to use an existing rebase action before
triggering this bubble_merge action. See workflow configuration below[1].

This commit introduces also a commit title and a commit message for the merge
bubble commit.

The commit message will contain the entire body of the comment which triggered
Squiddy.

Further information in the PR[2].

[1]
name: Bubble Merge
on:
  issue_comment:
    types: [created]
    branches-ignore:
      - master

jobs:
  bubble_merge:
    name: Bubble Merge
    if: github.event.issue.pull_request != '' && contains(github.event.comment.body, '@squiddy-bot merge')
    runs-on: ubuntu-latest
    steps:
      - name: Checkout the latest code
        uses: actions/checkout@main
        with:
          fetch-depth: 0
      - name: Automatic Rebase
        uses: cirrus-actions/rebase@1.3.1
        env:
          GITHUB_TOKEN: ${{ secrets.SQUIDDY_BOT_GITHUB_TOKEN }}
      - name: Acknowledge Failure
        if: ${{ failure() }}
        uses: actions/github-script@0.9.0
        with:
          github-token: ${{secrets.SQUIDDY_BOT_GITHUB_TOKEN}}
          script: |
            await github.issues.createComment({
              issue_number: context.issue.number,
              owner: context.repo.owner,
              repo: context.repo.repo,
              body: 'Rebase was not possible. Please rebase manually and try again.'
            })
      - uses: futurelearn/squiddy/actions/bubble_merge@mt-et-bubble-merge
        env:
          GITHUB_TOKEN: ${{ secrets.SQUIDDY_BOT_GITHUB_TOKEN }}
          GITHUB_EVENT: ${{ toJson(github.event) }}

[2]
https://github.com/futurelearn/squiddy/pull/3