Fix bug with posting same item multiple times
This is a pretty major bug in that the `item_id` method uses
`cards.checklists.first` to check the items, and it's extremely likely
that multiple checklists exist. This seems to be accidentally imported
from my testing, and I didn't catch it since we didn't have multiple
checklists.

Instead, use the `checklist` method which returns the checklist we're
interested in.