Remove Dependabot & Pull Request Checklist
I have replaced this code with Go running in a Lambda[1].

I enjoyed writing this but I found GitHub Actions too slow to run these
particular tasks, which had been a bugbear of mine when trying to
implement them. GitHub Actions takes a long time to start each action
because it has to fetch the dependencies and using Ruby meant that in
order to get it to be quicker, we did it using a pre-built Docker image.

This Docker image still had to be fetched each time the action ran, so
it was taking ~30s to run.

Using Lambda and a webhook, these "actions" now run in a few seconds,
which feels about right when adding a Trello URL to a pull request, and
see it immediately pop up on the Trello card. I think the performance
opens us up to do more complicated actions between Trello & GitHub.

While I have implemented these in a Lambda, I'm not neccessarily
convinced the bubble merge action is best suited to also moving to
Lambda. People seem pretty happy with it, and while it takes a minute or
so to run, it has to do stuff with a lot more data (ie cloning entire git
repositories for a rebase) rather than just send API calls between
different services.

I'm deleting all this code so there is no confusion between the two
repos.

[1] https://github.com/futurelearn/squiddy-go