Remove duplicate git client file
We're changing the git client file that is present in the bubble
merge action folder instead of this lib file.

We originally created a hard link to this file in the action folder
because Docker wasn't able to use a file outside of it's own folder.
See [1] for details.

It's a bit confusing to have two versions of the file around and
keep them in sync.

So let's remove the lib version and just use the one in the action
folder.

[1]: https://github.com/futurelearn/squiddy/commit/7ff33ceacd8a77a1bf9577be9ad444f03b534b98