Fetch CI status for a PR
Github considers the status of the last commit in a PR ("head commit")
as the status of a PR.

So we first need to fetch this head commit and then ask the API
what status it has.

We initially attempted to use the `/statuses` Github API endpoint for
status checks. However, this is discouraged [1] as checks performed
via Github Actions will not push "status" events so the list of statuses
will always come back empty. This is unfortunate since this particular
endpoint would return an overall status based on all the checks.

Instead it's recommended that we use the `/check_suites` endpoint [2]
which will return a list of **check_runs** (not "check_suites") for our
head commit [3], one for each github action.

These `check_runs` will each have a `status` and a `conclusion` field.

NB: One important thing to note is if the `conclusion` field exists,
`status` will automatically be set to `completed`. [4]

We've reflected this in upcoming tests.

Since we don't have an overall status for all checks returned for this
endpoint, we have to cycle through these `check_runs` to find any still
in progress or that have failed. If we can't find any (and they're all
marked as completed), then we go ahead with the merge.

Even though we're not making use of all possible statuses and conclusions
for a check_run, we've added a list of possible values for them as constants
for future reference.

[1]: https://docs.github.com/en/free-pro-team@latest/rest/reference/repos#get-the-combined-status-for-a-specific-reference
[2]: https://github.community/t/state-is-pending-statuses-are-missing-with-github-api-v3/14386/3
[3]: https://docs.github.com/en/free-pro-team@latest/rest/reference/checks#list-check-suites-for-a-git-reference
[4]: https://docs.github.com/en/free-pro-team@latest/rest/reference/checks#create-a-check-run