Disable CI checks
When we turned on Squiddy CI checks on the main Rails repo, we arrived
in a state where we were constantly rebasing since the repo has more
activity. This in turn would trigger CI checks which wouldn't allow us
to merge.

We've decided to turn the checks off so we don't block developers from
merging.