Update bubble_merge.yml
Currently when a comment is edited, github actions will not check to see if the edited comment contains "@squiddybot merge".

This means that if you make a typo and edit your comment, nothing will happen, which is a bit counterintuitive. 

This commit extends bubble merge to edited comments. If the comment does not contain "@squiddybot merge" the workflow is triggered, but the job is skipped, the same as when adding a new comment.