Migrate Bubble Merge workflow to a central reusable workflow
It is now possible to reference reusable workflows in
other private repositories owned by the same organization[1]

A central reusable workflow for Bubble Merge has been created
in the futurelearn/actions repo[2] and now we will reference
that from our other repos that want to use squiddy-bot to
perform bubble merges of pull requests.

By centralising the workflow, if any changes are required to it in
the future, we only need to update it in one place for it to apply
in all repos that use it.

The `branches-ignore` key has been removed as that is not valid
for `issue-comment` type triggers:

```
actionlint .github/workflows/bubble_merge.yml
.github/workflows/bubble_merge.yml:5:5: "branches-ignore" filter is not available for issue_comment event. it is only for push,
pull_request, pull_request_target, workflow_run events [events]
```

[1]https://github.blog/changelog/2022-12-14-github-actions-sharing-actions-and-reusable-workflows-from-private-repositories-is-now-ga/
[2]https://github.com/futurelearn/actions/commit/2e70636c43fb1a0e999489a3a7d84dec0864de84