Update actions/checkout
This dependency was relying on node v16, which github are deprecating
support for.

This will resolve the issue for one of the actions listed here

```
Node.js 16 actions are deprecated. Please update the following actions
to use Node.js 20: actions/checkout@v3

For more information see: https://github.blog/changelog/2023-09-22-github-actions-transitioning-from-node-16-to-node-20/.
```